Move convertDate to calculation module
